remove guestfish as a dependency

+++ build_disk.sh
@@ -1,34 +1,67 @@
 set -e
 
-rm -f disk.img
+# Clean up the state
 
-#wildcards are forbidden in guestfish, so we need to list the dirs in advance
 
-guestfish -N disk.img=disk:30M <<EOF
+cleanup () {
+	DEV=$(losetup | grep disk.img | cut -d ' ' -f 1)
+	if [ -z $DEV ]; then
+		return
+	fi
+	sudo losetup -d $DEV
+	sudo umount ./mnt
+	rm -f disk.img
+}
 
-part-init /dev/sda mbr
-part-add /dev/sda primary 2048 40960
-part-add /dev/sda primary 40961 -40
-part-set-bootable /dev/sda 1 true
+# initial clean up (I'm paranoid)
+cleanup
 
-mkfs vfat /dev/sda1
-mkfs ext2 /dev/sda2
+qemu-img create -f raw disk.img 40M
 
-mount /dev/sda2 /
-copy-in iso/home /
-copy-in iso/etc /
+# builidng the disk structure
+fdisk disk.img <<EOF
+n
+p
+1
+2048
+40960
+a
 
-mkdir /boot
-mount /dev/sda1 /boot
+n
+p
+2
+43008
+81919
 
-mkdir /boot/grub
-copy-in iso/boot/grub/grub.cfg /boot/grub
-
-grub-install / /dev/sda
+w
 EOF
 
-# if one day I get crazy and want to install grub2 manually
-#part-set-name /dev/sda 1 'EFI System'
-#part-set-gpt-type /dev/sda 1 c12a7328-f81f-11d2-ba4b-00a0c93ec93b
+# mounting shenanigans
+sudo losetup -fP disk.img
+DEV=$(losetup | grep disk.img | cut -d ' ' -f 1)
+if [ -z $DEV ]; then
+	echo "Failed to create loop device !"
+	exit 1
+fi
+P1=$DEV"p1"
+P2=$DEV"p2"
+echo "Device mounted on $DEV, parts are $P1 and $P2"
 
+# creating filesystems
+sudo mkfs.vfat $P1
+sudo mkfs.ext2 $P2
 
+# mounting, installing grub and copying config files/data files
+# boot partitions
+sudo mount $P1 ./mnt
+sudo grub-install --target=i386-pc --boot-directory=./mnt/ $DEV
+sudo cp ./disk/boot/grub/grub.cfg ./mnt/grub/
+sudo cp -r ./mojitos.elf ./mnt
+sudo umount ./mnt
+# root filesystem partition
+sudo mount $P2 ./mnt
+sudo cp -r ./disk/{etc,home} ./mnt
+sudo umount ./mnt
+
+# cleaning up
+cleanup
